+++ packages/components/src/index.ts
@@ -8,6 +8,7 @@ export { MinisPillCounter } from './components/pill-counter/pill-counter.js';
 export { MinisAlert } from './components/alert/alert.js';
 export { MinisMessage } from './components/message/message.js';
 export { MinisTag } from './components/tag/tag.js';
+export { MinisMenu } from './components/menu/menu.js';
 export { MinisNavigation } from './components/navigation/navigation.js';
 export { MinisNavigationItem } from './components/navigation/navigation-item.js';
 export { MinisCardGrid } from './components/card-grid/card-grid.js';